drt: Add nearby grid pin access strategy
The current pin access strategies fail on sky130_fd_sc_hs__xor3_1:B
where the pin is mostly blocked by an obstruction above.  This
change introduces another strategy tried if all the existing ones
fail.  We look at on-grid and 1/2-grid points with min width
of a pin shape for pin access.

The "iterations" were rolled up into a nested-for loop that was
implied by the previous code.