Force LC_ALL and LANG to C to avoid problem with localization.
In the Brazilian locale numbers have 0,5 format rather than the C 0.5.
This causes LEF parsing to error as it uses strtod which is locale
sensitive.

Strangely this problem only occurs when the GUI is initialized but I
set these envars globally avoid any further issues.  Qt locales must
be involved.

If we ever internationalize this may have to be revisited.

Fixes The-OpenROAD-Project/OpenROAD#1229

Signed-off-by: Matt Liberty <mliberty@eng.ucsd.edu>