Support hierarchical netlist for SPEF annotation
This is not full hierchical support but it does allow read_spef to
traverse the hierchy by name to find the leaf dbInst.  It doesn't
enable a full hierchy model in sta as odb isn't ready for that.

The dbNetwork will allow lookup of hierchical instance by name but
the iterators still operate flat.

Signed-off-by: Matt Liberty <mliberty@eng.ucsd.edu>