gpl: ensure the pin access is from the preferred routing direction
If you have a multi-layer pin the prior code would always use the top layer
as the routing target.  However the top layer may imply a wrong way
connection.  This change computes which edge the pin is on and finds the
highest right-way layer to connect to it (using the top layer if none
exists).

Signed-off-by: Matt Liberty <mliberty@eng.ucsd.edu>