drt: fix bug in METALWIDTHVIAMAP handling in rp
In prep_via2viaForbiddenLen_widthViaMap the wrong cut layer was being
used in the check and the second check had the wrong logic.  The result
was we disallowed stacking up and allowed stacking down erroneously.

I reduced this to one check as there is only a single cut layer
invovled.  I also refactored the code for readability.

Signed-off-by: Matt Liberty <mliberty@eng.ucsd.edu>