Fixes GF180MCU antenna violations
Fixes https://github.com/The-OpenROAD-Project/OpenROAD/issues/2516

The issue looks to be down to GF180MCU being simpler than we expect.

In SKY130 we have antenna rules that look like this.
```
  ANTENNAMODEL OXIDE1 ;
  ANTENNADIFFSIDEAREARATIO PWL ( ( 0 400 ) ( 0.0125 400 ) ( 0.0225 2609 ) ( 22.5 11600 ) ) ;

```

However, in GF180MCU we have rules that look like.
```
    ANTENNAMODEL OXIDE1 ;
    ANTENNADIFFSIDEAREARATIO 400 ;
    ANTENNAGATEPLUSDIFF 2 ;
```

There is only a single value instead of a linearly interpolated list of values.

```c++
bool AntennaChecker::antennaRatioDiffDependent(dbTechLayer* layer)
{
  if (layer->hasDefaultAntennaRule()) {
    const dbTechLayerAntennaRule* antenna_rule = layer->getDefaultAntennaRule();
    dbTechLayerAntennaRule::pwl_pair diffPAR = antenna_rule->getDiffPAR();
    dbTechLayerAntennaRule::pwl_pair diffPSR = antenna_rule->getDiffPSR();
    return diffPAR.indices.size() > 1 || diffPSR.indices.size() > 1;
  }
  return false;
}
```

The following code asserts that there must be at least two values for the rule to be diff dependent which is incorrect. When I changed the code to

```c++
bool AntennaChecker::antennaRatioDiffDependent(dbTechLayer* layer)
{
  if (layer->hasDefaultAntennaRule()) {
    const dbTechLayerAntennaRule* antenna_rule = layer->getDefaultAntennaRule();
    dbTechLayerAntennaRule::pwl_pair diffPAR = antenna_rule->getDiffPAR();
    dbTechLayerAntennaRule::pwl_pair diffPSR = antenna_rule->getDiffPSR();
    return diffPAR.indices.size() >= 1 || diffPSR.indices.size() >= 1;
  }
  return false;
}
```

OpenROAD fixed nearly all of the antennas in this design.

Signed-off-by: Ethan Mahintorabi <ethanmoon@google.com>