dpl: improve the instance on macro strategy
The current strategy for instance on macro is to push it the nearest
edge before starting the diamond search.  This works poorly if the
nearest edge has no nearby sites (eg a macro on the edge of the core
area where the nearest edge is the core edge).

Instead try the 'is_hopeless' strategy first which does look for a
site in the direction being moved to.  The old strategy is kept as
a fallback.

Signed-off-by: Matt Liberty <mliberty@eng.ucsd.edu>