gpl: remove early stopping code
The code is intended to avoid long waits on designs that won't converge.
However it stops too early on some designs involving large macros where
expanding the maze search window needs to continue until a net can be
routed around the macro.

This could be made more intelligent and restored in the future.

Signed-off-by: Matt Liberty <mliberty@eng.ucsd.edu>