drt/pa: Fix bounding box calculation in genPatterns_gc
genPatterns_gc miscalculates the bounding box used for gc, applying the
lower x coordinate to all other coordinates. If we are unlucky this
bounding box is huge and could contain over 100,000 objects. As a result
gc wastes a lot of time initialising all the objects within the bounding
box.

Fixing this significantly improves the start up time for the
ispd19_test10 test case.

Signed-off-by: Anton Blanchard <anton@linux.ibm.com>