Refactoring to enable a CTS unit test.
This splits all the remaining targets necessary to run a "librarified"
CTS unit test, that does not pull in the GUI and all openroad
singleton-based dependencies.

It runs a simple unit test, which doesn't look like much, but it starts
on a unit testing methodology where we can iterate on the algorithms
(e.g. CTS) in isolation.

Co-authored-by: Ethan Mahintorabi <ethanmoon@google.com>
Signed-off-by: Christopher D. Leary <cdleary@gmail.com>