Split GUI dependencies from dbSta definition.
This still does not split the OpenRoad observer object from the dbSta
class definition, so the file cannot be made private to the dbSta_lib
library at this point. A follow-on change intends to split the OpenRoad
observer dependency and make the dbSta.cc file not part of the
"application" target dbSta.

Co-authored-by: Ethan Mahintorabi <ethanmoon@google.com>
Signed-off-by: Christopher D. Leary <cdleary@gmail.com>