gui: reduce the rtree size by storing dbSBox instead of the cuts
For large multi-cut vias it is more efficient to store the dbSBox once
in the rtree rather than all the individual cuts.  Since this tends to
happen in the power grid this is applied only to special nets.

The tradeoff is that the rendering will have to do some extra work.
However the search is conducted in a smaller rtree which will help
to offset this.

Signed-off-by: Matt Liberty <mliberty@precisioninno.com>