dpl: Fixes out of bounds in grid paint
In some cases you can have a double height row
with only a single layer row below it instead of two.
In that case if you were painting a double height
cell then you would go out of bounds.

This commit adds a check to ensure, that does not happen.

Signed-off-by: Ethan Mahintorabi <ethanmoon@google.com>