rsz: Fix split length calculation for long wires
The `repairNetWire` function works on one wire segment at a time. To
detect an overly long wire, we compare `wire_length > max_length_`,
where `wire_length` is inclusive of previous segments.

When we request splitting, `split_length` is in terms of the current
segment only, so add a missing offset that accounts for the length
contributed from previous segments.

Signed-off-by: Martin Povišer <povik@cutebit.org>