rmp: Adds new logic cut library to remapper
This new logic remapper is step 2 of 3 to begin to do
restructuring in the loop of global placement. This new
logic library is extensible and can implement diferent
stratigies.

This version of the LogicExtractor takes a list of endpoints
and will walk backwards until it reaches a cell that ABC cannot
handle.

Signed-off-by: Ethan Mahintorabi <ethanmoon@google.com>