Create Nix Flake
Replacing `etc/shell.nix`, this is a [flake](https://nixos.wiki/wiki/Flakes) based on the OpenROAD Nix derivation used for OpenLane 2. With Nix installed with flakes enabled, getting an OpenROAD environment is as simple as typing `nix develop`.

Also fixes issue with versions of OpenROAD built in the dev environments failing to launch the GUI.

Signed-off-by: Mohamed Gaber <donn@efabless.com>