dbSta: in dbNetwork::parent move the non-hier case to dbInsts only
The complication with the previous change is that read_spef does want
to see the module hierarchy (it breaks up hierarchical names and
walks down the hierarchy).  The prior change prevented parent()
from working in that case.

Moved the check to just dbInst and not dbModInst to fix read_spef.

Signed-off-by: Matt Liberty <mliberty@precisioninno.com>