odb/cdl: Handle top-level pins wired to nets of different name
ATM the code seem to assume that for each top level pin, the
net name is equal to the pin name.

That's not always the case and this produces an incorect netlist.

To solve that, if a net is connected to a pin, we use the pin name
instead of the DB net name.

Signed-off-by: Sylvain Munaut <tnt@246tNt.com>