rmp: Adds backedge from ABC to OR
This PR adds a new function called LogicCut::InsertMappedAbcNetwork,
this function takes an Abc_Ntk_t* and reinserts it into the STA
netlist.

This function also removes the old logic cut and replaces it with
new instances derived from the Abc network.

Only thing left to do is optimize the abc network in some positive
way.

😎

Signed-off-by: Ethan Mahintorabi <ethanmoon@google.com>