mpl2: handle the track pattern more directly
If you have multiple track patterns (eg m2 in asap7) the current code
is wrong.  This is a partial solution in that it just uses the first
track pattern.  Ideally it should consider snapping to any of the
track patterns.  The cost will be snapping further than needed.

This "fixes" #6267.

Signed-off-by: Matt Liberty <mliberty@precisioninno.com>