Cleanup: Fix missing header in parsers.h and some clang-tidy issues.
The std::ifstream is used, but necessary header was missing.
Also `<cstring>` was missing for the strcmp().
While at it, remove superfluous headers.

Also: simplify a boolean return (there was a else after return
issue, and the whole expression could be written more concise and
readable anyway).

If there is a single statement in a block in a new line, add the
necessary curly braces.

Signed-off-by: Henner Zeller <h.zeller@acm.org>