Initial switch to bzlmod, moving most dependencies there.
Most dependencies are now pulled from bzlmod, but there are
a few remaining that are still pulled from bazel_rules_hdl,
fetched in WORKSPACE. Minimal change in that file for now, but
next step would be to prune this file and _only_ pull in the
dependencies only needed from there (such as lib ABC).

There is one issue with boost.beast in which the provided library
does not seem to provide all link targets and fails to link. Deferred
to later, for now, use the old `@boost//:beast` (from bazel_rules_hdl)
in `src/utl/BUILD`.

Signed-off-by: Henner Zeller <h.zeller@acm.org>