Fix unnecessary const qualified return values. Make some cons refs.
Returning a _value_ 'const' does not actually have any effect, it
just makes the reader wonder why that is (the compiler spits
out a ton of warnings about this).

Looking at this it turns out that some expensive types such as
strings, vectors, or maps have been returned by value which is
expensive.

Replace these with const reference for simple things (such as
strings), but left containers to be returned by value as it
requires first to analyse if the call sites, while iterating
through it, don't add more components, invalidating said
iterators. Added comment to that effect.

Signed-off-by: Henner Zeller <h.zeller@acm.org>