bazel: don't allow empty glob() patterns.
Patterns that don't result in anything are certainly accidental,
so have bazel double-check that we don't use such patterns.
With bazel 8, this will be the default, so this prepares us for
that as well.

Required to update a bunch of boost minor packaged versions as earlier
ones had that exact problem which are now reported.

Signed-off-by: Henner Zeller <hzeller@google.com>