bazel: make environment compatible with more operating systems.
When running the compilation on NixOS, it is not possible currently
to use the llvm_toolchain mentioned in the MODULE.bazel, but necessary
to run the local toolchain (`clangStdenv`). That in turn is strict and
only allows to compile C files when invoked with `clang` and
requires `clang++` or `clang -xc++` to compile C++ files - unlike a typical clang
installation that attempts to auto-detect. But Bazel is lazy and
invokes c++ compilation with `clang` ...

Another issue: the latest rules_cc version has hardcoded paths to
`/bin/bash` while the correct way to use that on Posix systems is to
call it via `/usr/bin/env`.

So, this PR solves both these issues:

  * Add `-xc++` to the compiler invocation for cxx compilation.

  * Use a rules_cc that has the fix to use /usr/bin/env
    (https://github.com/bazelbuild/rules_cc/pull/306), but that
    is not released yet on BCR (once 0.1.2 is out we can probably
    use that directly).

Signed-off-by: Henner Zeller <hzeller@google.com>