grt: Update layer lengths utility
The `grt::route_layer_lengths` method is used in OR flow scripts to
correlate the final parasitic extraction with wire length on individual
layers. This way one can extract empirical coefficients for
`set_wire_rc`.

Enhance this method to include the number of via crossings, and remove
"last-mile pin connection patching" which was introducing duplicate
vias.

Signed-off-by: Martin Povišer <povik@cutebit.org>