rsz: Fix ideal clock handling in driver delay modeling
In buffering, when modeling the driver delay change due to a changed
load, if the driver arc goes from a clock pin to a register output pin,
make the change to consider an idealized clock arrival on the clock pin
instead of looking up the arrival value on the vertex.

For arcs going from an ideal clock pin we already override the slew
(via the helper `graph_delay_calc_->edgeFromSlew()`), but an arrival
override was missing until now.

Signed-off-by: Martin Povišer <povik@cutebit.org>