histogram: rename 'Refresh filters' to 'Update'
This button is useful when issuing interactive set_false_path
commands or other commands that affect the timing reporting.

Normally when using dropdowns in Endpoint Slack Histogram,
then there's no need to click this button. Unsure what the
what 'Refresh filters' is referring to or what the use
case is other than to refresh the histogram based on out
of bound(not through GUI buttons) changes to the timing model.

Signed-off-by: Øyvind Harboe <oyvind.harboe@zylin.com>