odb: check for invalid strings being converted to enums in TCL
Swig is confusing in that it has both 'typecheck' and in' typemaps.
'typecheck' is run before overload resolution to determine if a string
belongs to a specific type.  After resolution 'in' is used to convert.

Confusingly if there is no overload to resolve then 'typecheck' is
skipped.  The previous code assumed 'typecheck' would always check and
the 'in' didn't need to also check.

Signed-off-by: Matt Liberty <mliberty@precisioninno.com>