bazel: fix ambiguous header definitions.
It should be possible to identify exactly the library that
provides a particular header and its implementation.
Make sure ther is an exact 1:1 relationship between
header and the library it is provided by.

Found using bant, printing the headers and the libraries
they are part of and showing the duplicates:

```
bant -q lib-headers | awk '{print $1}' | sort | uniq -cd
```

Signed-off-by: Henner Zeller <h.zeller@acm.org>