mpl util.h: use less ambiguous name.
The "util.h" has a very generic name and it is included not
relative to project root, but relative to current files, so relies
on the include search path.

Make this a more unique name and use mpl-util.h instead.

While at it, fix another "util.h" ambiguity in the rcx
header directory. Here, we can dismabiguate by adding the "rcx/"
prefix as it should be.

Signed-off-by: Henner Zeller <h.zeller@acm.org>