Remove using declaration that spilled a symbol into different namespace.
First step in fixing `using` declarations in header files. In this
case, the `odb::dbTransform` was using-declared in a different namespace
in a header file which then spilled this symbol into `drt::` namespace.
Unless in very special circumstances this is of course never desirable.

(using declaration was in src/drt/src/db/infra/frBox.h)

This is the first step in cleaning up `using` declarations that
spill symbols globally into different namespaces.

Signed-off-by: Henner Zeller <h.zeller@acm.org>