gpl: Improved options parsing (example template)
Demonstrated on the global_place command.  This moves the options to a
C++ struct (PlaceOptions) rather than a large number of setter methods
on the Replace class.  The options are passed through as a const
object rather than being class members.

The keys & flags from tcl are parsed in the .i file to allow the
defaults to come from the C++ struct rather than repeating them in
tcl.  Some helpers are provided in options.i and utl/options.h.

In python the need for gpl_aux is removed as commands can instantiate
and configure the PlaceOptions struct directly.

In C++ the PlaceOptions instantiation can use designated-initializer
syntax for brevity.

Signed-off-by: Matt Liberty <mliberty@precisioninno.com>