test/orfs/gcd: default options
The goal of the gcd test is to quickly run through the flow.

Even the future eqy tests work without messing with default options, so
set everything to default to reduce cognitive load of what this test
is testing.

Tests run quickly too for gcd, so no reason to mess with options.

Signed-off-by: Øyvind Harboe <oyvind.harboe@zylin.com>